Implement basic multiplexing
Remove server property from individual grpc services and implement it
instead as a singleton inside grpc. No attempt yet to make it thread
safe.

Existing tests are passing.

+++ api/grpcserver/grpcserver_test.go
@@ -279,9 +279,9 @@ func TestNewServersConfig(t *testing.T) {
 		port1, &networkMock, txAPI, nil, nil)
 
 	jsonService := NewJSONHTTPServer(port2, port1)
-	require.Equal(t, uint(port2), jsonService.Port, "Expected same port")
-	require.Equal(t, uint(port1), jsonService.GrpcPort, "Expected same port")
-	require.Equal(t, uint(port1), grpcService.Port(), "Expected same port")
+	require.Equal(t, port2, jsonService.Port, "Expected same port")
+	require.Equal(t, port1, jsonService.GrpcPort, "Expected same port")
+	require.Equal(t, port1, grpcService.Port(), "Expected same port")
 }
 
 func TestNodeService(t *testing.T) {
